Add files via upload
Regression analysis is a statistical method used to examine the relationship between a dependent variable and one or more independent variables. It is commonly used to predict or estimate the value of the dependent variable based on the values of the independent variables.

The main goal of regression analysis is to find the best-fitting line or curve that represents the relationship between the variables. This line or curve is called the regression equation. The equation can be used to make predictions or estimate the value of the dependent variable for a given set of independent variables.

There are different types of regression analysis, including simple linear regression, multiple linear regression, and nonlinear regression. Simple linear regression involves only one independent variable, while multiple linear regression involves two or more independent variables. Nonlinear regression is used when the relationship between the variables is not linear and requires a more complex equation.

Regression analysis also provides information about the strength and direction of the relationship between the variables. The coefficient of determination (R-squared) is often used to measure the proportion of the variance in the dependent variable that can be explained by the independent variables.

Regression analysis is widely used in various fields, including economics, finance, social sciences, and healthcare. It helps researchers and analysts understand the relationship between variables and make predictions or forecasts based on the data.